fix(mocks): update return signature for ready()
Upon upgrading to the latest of several libraries (mostly Ionic and Angular),
I started getting this error:

typescript: test-config/mocks-ionic.ts, line: 7
Argument of type '"READY"' is not assignable to parameter of type
'{ String: any; } | PromiseLike<{ String: any; }>'.

The signature as it was looked for the promise to resolve with an object
having a property called "String" which contained anything.

The actual signature found here:
https://github.com/ionic-team/ionic/blob/master/src/platform/platform.ts
calls for the promise to be resolved with a string.